Added Kick button in Teams GUI, removed all getColorByID() stuff. Team messages now are text components, so they support urls, special formatting etc. Started moving team messages to Teams GUI (Player left team, etc). Fixed notifications not working if no text is added. Removed min and max from IConfigValue, added addInfo() instead. Added CentredTextButton, removed Converter, InfoPageTheme (replaced with config)

+++ src/main/api/com/feed_the_beast/ftbl/api/IFTBLibClientRegistry.java
@@ -14,11 +14,11 @@ public interface IFTBLibClientRegistry
 {
     void addClientConfig(IConfigKey key, IConfigValue value);
 
-    default ConfigKey addClientConfig(String mod, String id, IConfigValue value, byte... flags)
+    default ConfigKey addClientConfig(String mod, String id, IConfigValue value, int... flags)
     {
         ConfigKey key = new ConfigKey(mod.toLowerCase() + '.' + id, value.copy());
 
-        for(byte b : flags)
+        for(int b : flags)
         {
             key.addFlags(b);
         }
