+++ src/main/lib/com/feed_the_beast/ftbl/lib/config/SimpleConfigKey.java
@@ -19,7 +19,7 @@ public class SimpleConfigKey extends FinalIDObject implements IConfigKey
     }
 
     @Override
-    public byte getFlags()
+    public int getFlags()
     {
         return 0;
     }
